added separate page for admin

+++ crud/views.py
@@ -19,15 +19,17 @@ from django.contrib.auth.decorators import login_required
 from django.db.models import Q, Count, Avg
 from .filters import PostFilter
 
-
+# Функция для отображения списка врачей и фильтрации по специальности, а также вывода средней оценки
 def index(request):
     if request.method == 'GET':
         query = request.GET.get('q')
+        #Поиск по специальности
         if query:
             results = Post.objects.filter(Q(last_name__icontains=query) | Q(speciality__icontains=query)).annotate(
                 num_comments=Count('comments')).order_by('-experience')
         else:
             results = Post.objects.all().annotate(num_comments=Count('comments')).order_by('-experience')
+        #Подсчет средней оценки анкет
         for post in results:
             comments = post.comments.filter(active=True)
             tone_scores = []
@@ -70,7 +72,10 @@ def index2(request):
     return render(request, 'crud/index2.html', {'results': results,
                                                 'filter': post_filter})
 
+#Функция для создания записи
 def create(request):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     form = PostForm()
     if request.GET.get('search_user'):
         search_query = request.GET.get('search_user')
@@ -152,9 +157,10 @@ import googletrans
 from googletrans import *
 
 
+#Функция для просмотра записи, добавления комментария, его перевода, определения тональности и сохранения в БД
 def view(request, id):
     post = get_object_or_404(Post, id=id)
-    comments = post.comments.filter(active=True)
+    comments = post.comments.filter()
     new_comment = None
     tone_scores = []
     comment_form = CommentForm(user=request.user)
@@ -165,6 +171,7 @@ def view(request, id):
                              'Вы не можете оставить отзыв, пока не записались на прием к врачу.')
         else:
             comment_form = CommentForm(data=request.POST, user=request.user)
+            #Определение языка комментария
             if comment_form.is_valid():
                 new_comment = comment_form.save(commit=False)
                 new_comment.post = post
@@ -176,6 +183,7 @@ def view(request, id):
                     else:
                         translated_comment.translated_body = translator.translate(new_comment.body).text
                     print(translated_comment.translated_body)
+                    #Определение тональности комментария
                     blob = TextBlob(translated_comment.translated_body)
                     tone_score = blob.sentiment.polarity
                     print(tone_score)
@@ -217,22 +225,32 @@ def view(request, id):
                                               "avg_tone_score": avg_tone_score,
                                               "score": score})
 
-
+#Функция для анкеты
 @login_required(login_url='/login/')
 def delete(request, id):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     Post.objects.get(id=id).delete()
     messages.info(request, 'Data Deleted successful')
     return redirect('crud/index')
 
-
+#Функция для редактирования анкеты
 @login_required(login_url='/login/')
 def edit(request, id):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     post = Post.objects.get(id=id)
     return render(request, 'crud/edit.html', {"post": post})
 
-
+#Функция для изменений анкеты
 @login_required(login_url='/login/')
 def update(request, id):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     post = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, instance=post)
     if request.method == 'POST':
@@ -246,11 +264,17 @@ def update(request, id):
 
     return render(request, 'crud/edit.html', context)
 
+def is_not_doctor(user):
+    return not user.groups.filter(name='Доктор').exists()
+
 
 from datetime import datetime, time
+from django.contrib.auth.decorators import user_passes_test
+from django.contrib.auth.decorators import login_required
 
-
+#Функция для записи на прием к врачу
 @login_required(login_url='/login/')
+@user_passes_test(is_not_doctor, login_url='profile2')
 def book_appointment(request, post_id):
     post = get_object_or_404(Post, id=post_id)
     form = AppointmentForm()
@@ -262,6 +286,7 @@ def book_appointment(request, post_id):
             time_slot = datetime.strptime(time_slot_str, '%H:%M').time()
 
             appointment_datetime = datetime.combine(date, time_slot)
+            #Условия для записи на прием
             if appointment_datetime <= datetime.now():
                 messages.warning(request, 'Выбранный временной интервал уже прошел. Пожалуйста, выберите другой слот.')
 
@@ -280,9 +305,11 @@ def book_appointment(request, post_id):
 
     return render(request, 'book_appointment.html', {'post': post, 'form': form})
 
-
+#Функция для отмены записи на прием к врачу
 @login_required
 def cancel_appointment(request, appointment_id):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     appointment = get_object_or_404(Zapis, id=appointment_id)
 
     if appointment.user != request.user and appointment.post.user != request.user:
@@ -296,9 +323,11 @@ def cancel_appointment(request, appointment_id):
 
     return render(request, 'cancel_appointment.html', {'appointment': appointment})
 
-
+#Функция для отмены записи от имени врача
 @login_required
 def reject_selected_appointments(request):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     if request.method == 'POST':
         if 'reject_selected' in request.POST:
             selected_appointments_ids = request.POST.getlist('appointment_id')
@@ -307,18 +336,75 @@ def reject_selected_appointments(request):
                 appointment.delete()
     return redirect('profile2')
 
+
+from datetime import datetime
 @login_required
 def reject_selected_appointments(request):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
     if request.method == 'POST':
         if 'reject_selected' in request.POST:
             selected_appointments_ids = request.POST.getlist('appointment_id')
             selected_appointments = Zapis.objects.filter(id__in=selected_appointments_ids)
+
+            # Проверка времени записи
+            current_time = datetime.now().time()
             for appointment in selected_appointments:
+                appointment_time = datetime.strptime(appointment.time_slot, '%H:%M').time()
+                if appointment.date < datetime.now().date() or (
+                        appointment.date == datetime.now().date() and appointment_time <= current_time):
+                    # Время записи уже прошло, пропустить отмену
+                    continue
                 appointment.delete()
+
             messages.success(request, 'Selected appointments rejected successfully.')
     return redirect('profile2')
 
 
+from django.contrib.auth.decorators import login_required, user_passes_test
+
+def is_admin(user):
+    return user.is_superuser
+
+@login_required(login_url='/login/')
+def admin(request):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
+
+    query = request.GET.get('q')
+    if query:
+        posts = Post.objects.filter(
+            Q(id__icontains=query) |
+            Q(first_name__icontains=query) |
+            Q(last_name__icontains=query)
+        )
+    else:
+        posts = Post.objects.all()
+    return render(request, 'crud/admin.html', {'posts': posts})
+
+@login_required(login_url='/login/')
+def auth(request):
+    return render(request, 'crud/auth.html')
 
 
+from .models import Comment
 
+@login_required(login_url='/login/')
+def comment_approval(request):
+    if not request.user.is_superuser:
+        return redirect('crud/auth')
+    comments = Comment.objects.order_by('-active', 'created_on')
+    if request.method == 'POST':
+        comment_id = request.POST.get('comment_id')
+        action = request.POST.get('action')
+        comment = get_object_or_404(Comment, id=comment_id)
+        if action == 'approve':
+            comment.active = True
+            comment.save()
+            return redirect('crud/admin_com')
+
+        if action == 'delete':
+            comment.delete()
+            return redirect('crud/admin_com')
+
+    return render(request, 'crud/admin_com.html', {'comments': comments})
